requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
schedule==1.2.1
pytz==2023.3
python-dateutil==2.8.2 